# Solver: Enforce dependencies on executables (fixes #4781).