… cabal-install/changelog
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/DSL.hs
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/DSL/TestCaseUtils.hs
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/MemoryUsage.hs
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/QuickCheck.hs
… cabal-install/tests/UnitTests/Distribution/Solver/Modular/Solver.hs
… cabal-testsuite/PackageTests/InternalVersions/BuildDependsBad/setup.cabal.out