@23Skidoo 2017-11-12
This all sounds absolutely great, but can you please add a short note to the changelog?