@grayjay 2017-11-12
Thanks.  I updated the changelog and added a few more tests.