@kosmikus 2017-11-22
Looks fantastic. Thanks a lot @grayjay . @23Skidoo Please merge.